Oops 2

+++ Chowdren/base/common.cpp
@@ -1573,7 +1573,7 @@ int Frame::get_loop_index(const std::string & name)
 
 void Frame::reset()
 {
-    std::cout << "Reset frame: " << (unsigned int)this << std::endl;
+    std::cout << "Reset frame: " << (unsigned long)this << std::endl;
 
     ObjectList::iterator it;
     for (unsigned int i = 0; i < MAX_OBJECT_ID; i++) {
